Remove risk schemas wrapper
